+++ packages/isolate-daemon/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @ricsam/isolate-daemon
 
+## 0.1.31
+
+### Patch Changes
+
+- Updated dependencies
+  - @ricsam/isolate-protocol@0.1.23
+  - @ricsam/isolate-playwright@0.1.23
+  - @ricsam/isolate-runtime@0.1.30
+
 ## 0.1.30
 
 ### Patch Changes