+++ packages/playwright/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ricsam/isolate-playwright
 
+## 0.1.23
+
+### Patch Changes
+
+- improve logging
+- Updated dependencies
+  - @ricsam/isolate-protocol@0.1.23
+
 ## 0.1.22
 
 ### Patch Changes